# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TASK MANAGER API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "b91575f479176230541020cbbe70d4d3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Sarvarbek"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1525152162"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "fullname": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "phone": "166632623",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "a31bffef3b64c1e19862cec1d22db401",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "1525152162"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "fullname": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get": {
            "get": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "78ab02b03e430325937c4d9613a11f02",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 9,
                                                "fullname": "Sarvarbek",
                                                "avatar": null,
                                                "age": null,
                                                "address": null,
                                                "token": "F0B0Jj9a5ZOrd5kAu2Dsgr6xx6S1XK",
                                                "password": "$2y$12$2ShDf.Ev8XavVZI7.PDcYOpkfj9DjfL3SQGrqVIdm.P9QIa0pZCPW",
                                                "phone": "994630613",
                                                "group_id": null,
                                                "course_id": null,
                                                "created_at": "2024-01-13T09:05:53.000000Z"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}